# swagger.yml
swagger: "2.0"
info:
  title: Moku Moku Cards & Decks API
  description: Flash cards learning platform's Cards & Decks CRUD micro-service API documention.
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - https
definitions:
  Card:
    type: object
    properties:
      id:
        type: string
      front:
        type: string
      back:
        type: string
      image:
        type: string
...
  RestErr:
    type: object
    properties:
      message:
        type: string
      status:
        type: integer
        format: int
      error:
        type: string
parameters:
  access_token:
    name: access_token
    type: integer
    in: header
    required: true
paths:
...
  /card:
    post:
      tags:
        - card
      consumes:
        - application/json
        - multipart/form-data
      operationId: postCard
      description: Create a new card
      parameters:
        - name: new_card_data
          in: body
          required: true
          schema:
            allOf:
              - $ref: '#/definitions/Card'
            type: object
            properties:
              id:
                type: string
                readOnly: true
        - name: file
          in: formData
          type: file
          description: Card image
          x-mimetype: image/jpeg
      responses:
        '200':
          description: A new card has been created
          schema:
            $ref: '#/definitions/Card'
        '400':
          description: Invalid json body
          schema:
            $ref: '#/definitions/RestErr'
        '500':
          description: Something failed while creating a card
          schema:
            $ref: '#/definitions/RestErr'
...
  /card/pics/{pic_hash}:
    parameters:
      - $ref: '#/parameters/access_token'
      - in: path
        type: string
        format: string
        name: pic_hash
        required: true
        description: The relative image path (image) from the Card DTO
    get:
      tags:
        - card
      produces:
        - image/png
        - image/jpeg
      operationId: GetCardPicture
      description: Retrieve the picture of a card
      responses:
        '200':
          description: Card picture retrieved correctly
        '404':
          description: Card picture not found